Refactor hub related stuff under curated_hub
